Move CreateRecord to the recordstore pkg
