+++ internal/pkg/service/buffer/store/recordstore/record.go
@@ -1,4 +1,4 @@
-package configstore
+package recordstore
 
 import (
 	"bytes"
